Change printing of dict syntax to match parser
Change printing for operations to specify how the operation is printed
in relation to its list of operands: either as a prefix, infix or
circumfix (alternating between each operand, starting with an operand).
Circumfix is only used by dict get and set for now.
It's possible some precedence ordering may have been slightly altered by
this change, but I have not had the chance yet to look through
carefully, so this is something to watch for.
Handles dict case of #66.