Finish remaining transferBgp code?
Implement rest of transferBgp behavior for hijack.
Might be some weird code that can be simplified or fixed: we also assume
there exists a b variable which has bgpType, which might be risky.
Change "blocked" to "leak" in cfg to be more obvious what's going on.
Making leak take effect is still to do.