(incomplete) update to benchmark tabulate by trial
Begin to change tabulate_sp_bench.py to distinguish separate trials and
return data in an easier-to-transfer format for use with Excel/GSheets.
For now, add an argument to switch off averaging in mean_float_dict:
eventually, we'd prefer to default to no averaging with distinct
operations, giving the maximum amount of visibility to the user to
inspect how multiple runs differed.