nvgen: handle hijack leaks and group comment
Allow leaking to be turned on or off for nvgen's hijack generator. The
last spine-to-hijacker edge will be changed to leak the route when
leaking is on.
Also add code so that a (slightly simpler) comment is added back into
nvgen-generated networks that includes the group information.
Since hijacking is only implemented for fattree networks at the moment,
the hijack will designate the hijacker node as part of the spine (core).