Add benchmark playing with symbolic dest
Add a new fat4Pol benchmark where d is now a symbolic IP address.
Leaving the rest of the policy as-is, we change the assertion to instead
say, either every node has a path to the destination, or every node has
a non-BGP route.
The second disjunct is for the case when d does not conform to an
address such that init generates a BGP route at a node: in such network
states, some nodes may have static or connected routes to d, but nobody
shares BGP announcements to the destination.
It would be good to check if an assertion like this can be partitioned.