Change down to option[tnode] in maintenance bench
Change down to be a symbolic option[tnode] instead of bare tnode in the
maintenance benchmarks.
This is primarily useful for considering partitioned versions of the
benchmarks.
With the old type of down, the benchmarks considered the cases where
exactly one node was down at a given time in the monolithic benchmark.
By changing the type to an option, the monolithic benchmarks now cover
states where zero or one node is down, which clearly overapproximates
the previous case.
This then also allows a partitioned version of fatXPolMaintenance where
each partition has an option[tnode] down variable it can independently
consider: this leads to a further overapproximation of the set of
solutions, where we consider states where up to N nodes may be down at
once, where N is the number of partitions.
The next step is of course to actually verify that our existing theory
is sufficient to prove the property in question for partitions where up
to N nodes may be down at once.