Hack aggressive interpreter in transform_assert
The interpreter sometimes reduces expressions in the assert conjunct so
well that they disappear entirely, i.e. become true and can be
eliminated.
Unfortunately, this can wind up leaving us with an expression which we
had never intended to evaluate, since we would have dropped that
conjunct when remapping conjuncts.
This is demonstrated by the node-eq.nv file.
To try and handle this, we add a hack case to transform_assert.

This hack may not work in all cases, and could potentially cause bad
behaviour! It needs to be changed for something more stable, like a
process which safely remaps expressions and then only drops the right
number of conjuncts from the expression.