Add z3 timeout for smt solve
Add a soft timeout per Z3 query for the Kirigami and Classic SMT solve
procedures.
The soft timeout is given in seconds and cancels the query if it
surpasses the time. Unknown is reported instead, with :reason-unknown
being "canceled".
We still need a handler for this case of returning unknown so that the
user can tell that the timeout worked as intended, esp. for the kirigami
benchmarks (where each query uses the same timeout).