Add a Z3 timeout smt_answer
Add an SMT answer for when Z3 times out, allowing us to distinguish SMT
returning unknown from it being cancelled. Reports the timeout to the
user.
Also, add some more print-out data to inform the user of what the result
was for different partitions, making it easier to know which partitions
succeeded, failed or timed out.