Drop find_input_symbolics, use String.compare
find_input_symbolics required a kludge-y approach to collecting the
relevant hypotheses which led to a significant slowdown in encoding
time when searching the ConstantSet for input symbolics.
This obviates the need for that function by keeping track of all
relevant hyp variables for each edge in input_exp, and then using those
directly during SmtClassicEncoding.

Currently, this approach loses one feature of the old one, which is that
the solution properly identified the type of hypotheses in
counterexamples: instead, we are given the basic constituent elements
after all other transformations, but which are not converted back.
More work needs to be done to address this, possibly by re-introducing
the hypotheses (not the best option) so that they can make use of the
transformers, or alternatively adjusting the transformers to be able to
modify the hypothesis variables themselves by having transformers also
affect the partitioned_srp types (also somewhat ugly to tack on).

In either case, we mainly want a cheap way to look up the relevant
hypotheses for each input at encoding time. Storing all the necessary
information in the partitioned_srp is sensible, but it may be that we
can find a balance between presenting a readable solution and keeping
encoding cheap that is better to this one.