Reduce repeated calls to interp for predicates
Remove some of the duplicate cases where we call interp to get a
predicate by collecting all predicates initially in get_predicates and
then adding them in one go.
There are still some broader issues with how partitioning is done and
how information is stored: there are several points where a more
efficient method for storing information on the partitions would save us
intermediary recalculations where we must rediscover the cross edges or
the old nodes and so on.
That being said, it appears that #73 is now closed: we can address
partitioning time more fully next.