Re-add wellformedness checks for nodes and edges
For whatever reason (possibly partitioning?), Wellformed.check
has not been verifying that nodes and edges belong in the network: this
re-adds that functionality.
Since nodes and edges are remapped by Kirigami, this should not affect
the verification of those networks.
Originally, Wellformed.check also did not ensure that all edges
reference nodes less than the declared number of nodes: this adds a
function Wellformed.check_valid_edges that does so.