Fix generating full mesh topologies, add dest
Change the topology operation to add topology information to an existing
network.
A future addition can be to allow an "empty" network file to be given
for this command.
Full mesh networks were being generated with no 0 nodes by ocamlgraph,
so a custom rewrite was added to Topologies.construct.
A function dest_decl was added to define a dest variable from the given
cfg.destination.