Add -fragments flag to Kirigami
Add a -fragments flag to Kirigami to allow users to ask for only some
partitions to be checked. The fragments string should be written as a
comma-separated sequence of "m-n" or "n" substrings, where m and n are
integers.
Any fragment which is not listed will be skipped during partitioning.