Change encode_solve to use VertexMap for labels
Change the type used in encoding so we can gradually ascend to the point
where we only manipulate vertices and can change the underlying type.
An important consideration here is that the arrays encoding may be
faster than using AdjGraph: if so, it could make more sense to simply
store an additional data structure which maps names to nodes, and retain
a simple internal representation of 0..n.
This would allow names to be arbitrary within the program body, and used
by the interpreter, but we would still need to do some form of rewriting
when partitioning to reassign each name's node index.