Replace "bare" use of nodes/edges w AdjGraph.t
There should no longer be as many places in the code where a vertex is
treated as a bare int: instead we should always defer to the
implementation in AdjGraph.ml.
This involves also replacing get_nodes/get_edges uses with get_graph.
While get_graph constructs the graph each time, which may be slightly
more work, this ensures for now that we don't try manipulating bare
values quite as much.
There may still be places in printing where nodes/edges are treated as
integers: those need to be fixed next.
As well, we still have a DNodes declaration that is an int, rather than
a list of Vertex.t.