A lot more cuts and cleanup
Simplify definition of partitioned_srp now that we don't need to do
remapping since nodes are stored as a list.
Delete OCamlUtils.list_seq and use List.init instead.
Kirigami code most likely does not work correctly anymore: this needs to
be addressed by fixing the assert transformations, which get screwy.
We also may need to still check for situations where we ask for a
particular node that is not in the network anymore, e.g.
```
assert (sol[0n])
```
but for now, we're continuing to assume a particular type of assert
statement (the foldNodes construct), so Kirigami can go ahead and try to
get that one fixed and we can handle the rest later.