Fix mapping of label_vars for mono, not kirigami
Some aspect of assigning label_vars is currently broken, where kirigami
or another part of the network makes assumptions about the order they
are given in when encoding to SMT and reporting results.
That being said, it works for the monolithic approach now: before, we
used an array and indexed by node name, which meant that we could
generate 3 variables for a network with nodes (1,2,3) and then try and
access label_vars.(3), leading to an array overflow.

Now, we instead need to find a way to use the mask representing which
nodes have been kept and which have been cut in order to select with
label_vars to use for what during encode_kirigami_solve.