Fix bug in kirigami encode label_vars
Bug was caused by some issues with reversing the order of nodes when
distinguishing the cut and kept variables of the network during
encode_kirigami_solve.
The code is now fixed to handle this correctly. We also add
documentation and notes about how map unrolling's implementation will
construct the maplist according to how the nodes and edges are returned
by AdjGraph, which can depend on the underlying graph implementation and
the implementation of AdjGraph.Vertex and AdjGraph.Edge.
Clarify how the partitioned_srp cut_mask field works, and mark down that
it contains the nodes in reverse order, assuming integer nodes.