Simplify trans_map and trans_input_map fns
The SmtClassicEncoding uses EdgeMaps for trans_map and trans_input_map,
but it can instead use plain VertexMaps and save us the extra step of
specifying which end of the edge we want to use.
These changes bring up the question of whether it would make sense to
use a labelled graph with these variables stored on the vertex labels.